pyproject.toml
[tool.poetry]
name = "channel-message-forwarder"
version = "0.1.0"
description = ""
authors = ["Hundera Awoke <hunderaweke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
python-decouple = "^3.8"
pymongo = "^4.6.2"
loguru = "^0.7.2"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
